% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census.R
\name{get_census}
\alias{get_census}
\title{get_census}
\usage{
get_census(landscape, year = 2010, spatial = FALSE, level = "block_group",
  api_key = Sys.getenv("CENSUS_API_KEY"), delete_zips = FALSE)
}
\arguments{
\item{landscape}{A spatial polygon of class "sf" (see package [sf]). Note: the object must be projected (i.e. has a coordinate reference system and must overlap a USA census area)}

\item{year}{Which census year do you want to use? In format YYYY.  Currently the package supports census years 2000, and 2010.}

\item{spatial}{keep the spatial data? Default = FALSE}

\item{level}{for 2000 level == "block_group" for 2010 choose level = "block" or level = "block_group"}

\item{api_key}{Your Census API key. Obtain one at http://api.census.gov/data/key_signup.html}

\item{delete_zips}{Important:  if set to TRUE all zip files in the working directory will be deleted. Each time you run get_census one or more zip files will be added to your working directory.  After processing these can be deleted.  The default is delete_zips=FALSE.  If you are unsure about this don't change it.}
}
\description{
This function takes an input landscape, a user-defined land area as an
[sf] polygon and returns the total population for that area for a given USA decadal census.
The input polygon must be projected and must overlap a USA census area
Data are returned for years 2000, or 2010; The default is 2010
For 2000 only the census block group data are available.  For 2010 the block level information
is available but can take a long time to download for larger polygons.  By default the function
downloads the block group level for 2010 but you can request the more detailed block data.
}
\examples{
get_census()
}
