% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goatscape.R
\name{goatscape}
\alias{goatscape}
\title{Summarize several landscape datasets for a given landscape polygon}
\description{
This package provides tools to summarize numerous landscape datasets for a 
given landscape polygon and returns the data in a tidy \code{sf} object.
}
