% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_nlcd.R
\name{gs_nlcd}
\alias{gs_nlcd}
\title{gs_nlcd}
\usage{
gs_nlcd(landscape, spatial = TRUE, label = NA, year = 2011,
  dataset = "landcover", raw.dir = "./nlcd_data/raw",
  extraction.dir = "./nlcd_data/clipped",
  raster.options = c("COMPRESS=DEFLATE", "ZLEVEL=9", "INTERLEAVE=BAND"),
  force.redo = TRUE)
}
\arguments{
\item{landscape}{A spatial polygon of class "sf" (see package [sf]). Note: the object must be projected (i.e. has a coordinate reference system and must overlap the NLCD area)}

\item{spatial}{keep the spatial data? Default = TRUE}

\item{label}{A character string naming the study area. If no label supplied the function will use the name of the input landscape as the label.  A unique label for each input landscape is required if force.redo = FALSE.}

\item{year}{An integer representing the year of desired NLCD product.
Acceptable values are 2011 (default), 2006, and 2001.}

\item{dataset}{A character string representing type of the NLCD product.
Acceptable values are landcover' (default), 'impervious', and 'canopy'.}

\item{raw.dir}{A character string indicating where raw downloaded files should be put.
The directory will be created if missing. Defaults to './RAW/NED/'.}

\item{extraction.dir}{A character string indicating where the extracted and cropped DEM should be put.
The directory will be created if missing. Defaults to './EXTRACTIONS/NED/'.}

\item{raster.options}{a vector of options for raster::writeRaster.}

\item{force.redo}{If an extraction for this template and label already exists, should a new one be created?}
}
\description{
This function takes an input landscape, a user-defined land area as an
[sf] polygon and returns the NLCD landscape, canopy, or impervious cover for the polygon.
The input polygon must be projected and must overlap a USA NLCD area
Data are returned for years 2001, 2006, or 2011; The default is 2011
}
\examples{
gs_nlcd()
}
